% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bsc.R
\name{calc_bsc}
\alias{calc_bsc}
\title{Returns a list of bivariate spatial correlation statistics
according to Lee S. J Geograph Syst (2001) 3:369-385.}
\usage{
calc_bsc(feature1, feature2, sdata, conn_mat, assay = "SCT")
}
\arguments{
\item{feature1}{a gene symbol as in Seurat object}

\item{feature2}{another gene symbol as in Seurat object}

\item{sdata}{Seurat formatted spatial 10X scRNA expression data}

\item{conn_mat}{connectivity matrix (conn_mat) from make_conn_mat}

\item{assay}{default to 'SCT' but can be 'Spatial' (raw) or 'SCT'}
}
\value{
A list containing $r (Pearson's r), $r_sm (spatially smoothened r), $L_XX (clustering factor of feature1), $L_YY (clustering factor of feature2), $L_XY (L_X,Y bivariate spatial correlation from)
}
\description{
This function converts input temperatures in Fahrenheit to Celsius.
}
\examples{
devtools::install_github('satijalab/seurat-data')
SeuratData::InstallData("stxBrain")
brain <- SeuratData::LoadData("stxBrain", type = "anterior1")
brain <- Seurat::SCTransform(brain, assay="Spatial", verbose=T, method="poisson") # from Seurat
conn_mat <- make_conn_mat(brain) # make connectivity matrix
feature1 <- 'Gpr88'
feature2 <- 'Penk'
bsc = calc_bsc(feature1, feature2, brain, conn_mat)
print_line = sprintf("\%s\t\%s\t\%.3f\t\%.3f\t\%.3f\t\%.3f\t\%.3f",
  feature_i, feature_j, bsc$L_XX, bsc$L_YY, bsc$r_sm, bsc$r, bsc$L_XY)
print(print_line)
}
