% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_conn_mat.R
\name{make_conn_mat}
\alias{make_conn_mat}
\title{Make connectivity matrix}
\usage{
make_conn_mat(sdata, orig.ident = NULL, verbose = F)
}
\arguments{
\item{sdata}{Seurat formatted spatial 10X scRNA expression data}

\item{orig.ident}{A unique orig.ident for creating connectivity matrix.}

\item{verbose}{Print intermediate value checks as stderr()}
}
\value{
A list containing $barcodes_in_tissue (barcodes of spatial cells with 'tissue=TRUE'),
$nbarcodes_in_tissue (length of barcodes_in_tissue),
$C (raw connectivity/adjacency matrix),
$W (row-sum divided connectivity matrix),
$L_estimate_divR (L estimate, divided by Pearson's R (used later))
}
\description{
This function calculates the connectivity of a spatial cell to other cells.
}
\details{
Returns a list of spatial connectivity matrix variables (C, W)
according to Lee S. J Geograph Syst (2001) 3:369-385.
}
\examples{
#' # Refer to https://github.com/soymintc/scBSC
}
