% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bsc.R
\name{calc_bsc}
\alias{calc_bsc}
\title{Calculate Bivariate Spatial Correlation}
\usage{
calc_bsc(feature1, feature2, sdata, conn_mat, assay = "SCT")
}
\arguments{
\item{feature1}{a gene symbol as in Seurat object}

\item{feature2}{another gene symbol as in Seurat object}

\item{sdata}{Seurat formatted spatial 10X scRNA expression data}

\item{conn_mat}{connectivity matrix (conn_mat) from make_conn_mat}

\item{assay}{default to 'SCT' but can be 'Spatial' (raw) or 'SCT'}
}
\value{
A list containing $r (Pearson's r),
$r_sm (spatially smoothened r),
$L_XX (clustering factor of feature1),
$L_YY (clustering factor of feature2),
$L_XY (L_X,Y bivariate spatial correlation from)
}
\description{
This function converts input temperatures in Fahrenheit to Celsius.
}
\details{
Returns a list of bivariate spatial correlation statistics
according to Lee S. J Geograph Syst (2001) 3:369-385.
}
\examples{
# Refer to https://github.com/soymintc/scBSC
}
