% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_conn_mat.R
\name{make_conn_mat}
\alias{make_conn_mat}
\title{Returns a list of spatial connectivity matrix variables (C, W)
according to Lee S. J Geograph Syst (2001) 3:369-385.}
\usage{
make_conn_mat(sdata, verbose = F)
}
\arguments{
\item{sdata}{Seurat formatted spatial 10X scRNA expression data}

\item{verbose}{Print intermediate value checks as stderr()}
}
\value{
A list containing $barcodes_in_tissue (barcodes of spatial cells with 'tissue=TRUE'), $nbarcodes_in_tissue (length of barcodes_in_tissue), $C (raw connectivity/adjacency matrix), $W (row-sum divided connectivity matrix), $L_estimate_divR (L estimate, divided by Pearson's R (used later))
}
\description{
This function calculates the connectivity of a spatial cell to other cells.
}
\examples{
brain <- Seurat::SCTransform(brain, assay="Spatial", verbose=T, method="poisson") # from Seurat
conn_mat <- make_conn_mat(brain) # make connectivity matrix
feature1 <- 'Gpr88'
feature2 <- 'Penk'
bsc = calc_bsc(feature1, feature2, brain, conn_mat)
print_line = sprintf("\%s\t\%s\t\%.3f\t\%.3f\t\%.3f\t\%.3f\t\%.3f",
  feature_i, feature_j, bsc$L_XX, bsc$L_YY, bsc$r_sm, bsc$r, bsc$L_XY)
print(print_line)
}
